Move express helmet to gateway
