+++ deploy/helm/magda-dev.yml
@@ -20,6 +20,7 @@ gateway:
   service:
     type: LoadBalancer
   loadBalancerIP: "104.199.180.124"
+  cspReportUri: https://sdga.report-uri.com/r/d/csp/enforce
   auth:
     facebookClientId: "173073926555600"
     googleClientId: "275237095477-f7ej2gsvbl2alb8bcqcn7r5jk0ur719p.apps.googleusercontent.com"
@@ -84,7 +85,6 @@ search-api:
       cpu: 100m
       memory: 0
 web-server:
-  cspReportUri: https://sdga.report-uri.com/r/d/csp/enforce
   resources:
     requests:
       cpu: 100m
