+++ deploy/helm/magda/charts/web-server/templates/deployment.yaml
@@ -19,8 +19,7 @@ spec:
             "--config", "/etc/config/config.json",
             "--listenPort", "80",
             "--baseExternalUrl", {{ .Values.global.externalUrl | quote }},
-            "--registryApiBaseUrlInternal", "http://registry-api/v0",
-            "--cspReportUri", {{ .Values.cspReportUri | quote }}
+            "--registryApiBaseUrlInternal", "http://registry-api/v0"
         ]
         readinessProbe:
           httpGet:
