+++ magda-gateway/src/index.ts
@@ -4,6 +4,7 @@ import * as express from "express";
 import * as path from "path";
 import * as yargs from "yargs";
 import * as ejs from "ejs";
+import * as helmet from "helmet";
 
 import addJwtSecretFromEnvVar from "@magda/typescript-common/dist/session/addJwtSecretFromEnvVar";
 
@@ -106,6 +107,11 @@ const argv = addJwtSecretFromEnvVar(
             demand: true,
             default:
                 process.env.USER_ID || process.env.npm_package_config_userId
+        })
+        .option("cspReportUri", {
+            describe:
+                "The URI to send Content Security Policy violation reports to.",
+            type: "string"
         }).argv
 );
 
@@ -118,6 +124,52 @@ const authenticator = new Authenticator({
 // Create a new Express application.
 var app = express();
 app.disable("x-powered-by");
+app.use(
+    helmet({
+        hsts: {
+            maxAge: 31536000,
+            includeSubdomains: true,
+            preload: true
+        }
+    })
+);
+
+const cspDirectives = {
+    scriptSrc: [
+        "'self'",
+        "'unsafe-inline'", // for VWO until... we get rid of that? :(
+        "data:", // ditto
+        "browser-update.org",
+        "dev.visualwebsiteoptimizer.com",
+        "platform.twitter.com",
+        "www.googletagmanager.com",
+        "www.google-analytics.com",
+        "rum-static.pingdom.net",
+        "https://cdnjs.cloudflare.com/ajax/libs/rollbar.js/2.4.1/rollbar.min.js",
+        "https://tagmanager.google.com/debug",
+        "http://assets.zendesk.com/embeddable_framework/main.js", // zendesk
+        "https://assets.zendesk.com/embeddable_framework/main.js" // zendesk
+    ],
+    objectSrc: ["'none'"],
+    sandbox: [
+        "allow-scripts",
+        "allow-same-origin",
+        "allow-popups",
+        "allow-forms",
+        "allow-popups-to-escape-sandbox"
+    ]
+} as helmet.IHelmetContentSecurityPolicyDirectives;
+
+if (argv.cspReportUri) {
+    cspDirectives.reportUri = argv.cspReportUri;
+}
+
+app.use(
+    helmet.contentSecurityPolicy({
+        directives: cspDirectives,
+        browserSniff: false
+    })
+);
 
 const configuredCors = cors({
     origin: true,
