+++ magda-web-server/package.json
@@ -43,7 +43,6 @@
     "@magda/typescript-common": "^0.0.44-0",
     "@magda/web-client": "^0.0.44-0",
     "express": "^4.15.3",
-    "helmet": "^3.12.0",
     "morgan": "^1.9.0",
     "request": "2.85.0",
     "sitemap": "^1.13.0",
